feat: отработана отрисовка плоскости через упаковочный класс Plane
