+++ EngineCore/CMakeLists.txt
@@ -20,6 +20,7 @@ set(ENGINE_PRIVATE_INCLUDES
 	includes/EngineCore/Rendering/OpenGL/IndexBuffer.h
 	includes/EngineCore/Rendering/OpenGL/RendererOpenGL.hpp
 	includes/EngineCore/Rendering/OpenGL/Texture2D.hpp
+	includes/EngineCore/Rendering/OpenGL/ShaderContainer.hpp
 	
 	includes/EngineCore/Rendering/OpenGL/Primitives/PrimitiveObject.hpp
 	includes/EngineCore/Rendering/OpenGL/Primitives/Plane.hpp
@@ -37,6 +38,7 @@ set(ENGINE_PRIVATE_SOURCES
 	src/EngineCore/Rendering/OpenGL/IndexBuffer.cpp
 	src/EngineCore/Rendering/OpenGL/RendererOpenGL.cpp
 	src/EngineCore/Rendering/OpenGL/Texture2D.cpp
+	src/EngineCore/Rendering/OpenGL/ShaderContainer.cpp
 	
 	src/EngineCore/Rendering/OpenGL/Primitives/PrimitiveObject.cpp
 	src/EngineCore/Rendering/OpenGL/Primitives/Plane.cpp
